requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
qrcode[pil]==7.4.2
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
Werkzeug==3.0.4
WTForms==3.1.2
itsdangerous==2.2.0
email-validator==2.2.0
